removing LoginButton component and including logic in HeaderNavMenu
